requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.2
django-cors-headers==3.13.0
django-shopping-cart==0.1
gunicorn==20.1.0
Pillow==9.2.0
pycodestyle==2.9.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
whitenoise==6.2.0
